Herstructureer javascripts
