--- resources/assets/js/app.ts
@@ -1,11 +1,0 @@
-import './bootstrap';
-import './ctx';
-import './dragobject';
-import './formulier';
-import './forum';
-import './grafiek';
-import './logger';
-import './maalcie';
-import './main-menu';
-import './register-vue';
-import './router';
