+++ resources/assets/js/entry/app.ts
@@ -1,0 +1,6 @@
+import '../bootstrap';
+import '../dragobject';
+import '../logger';
+import '../main-menu';
+import '../register-vue';
+import '../router';
