+++ resources/assets/js/entry/extern-defer.ts
@@ -3,19 +3,19 @@
  */
 import axios from 'axios';
 import $ from 'jquery';
-import {registerBbContext, registerFormulierContext} from './context';
-import {init} from './ctx';
-import './fotoalbum/main';
+import {registerBbContext} from '../context';
+import {init} from '../ctx';
+import {route} from '../lib/util';
 
 require('lightbox2');
-require('./lib/external/jquery.markitup');
-require('jquery-ui/ui/widgets/tooltip');
+require('../lib/external/jquery.markitup');
 require('jquery-hoverintent');
 
 require('timeago');
 
-registerFormulierContext();
 registerBbContext();
+
+route('/fotoalbum', () => import(/* webpackChunkName: "fotoalbum" */'../page/fotoalbum'));
 
 declare global {
 	// Deze functie heeft geen type...
