+++ resources/assets/js/entry/fxclippy.ts
