+++ resources/assets/js/entry/fxminion.ts
