+++ resources/assets/js/entry/fxtrein.ts
