+++ resources/assets/js/entry/ledenmemory.ts
@@ -3,7 +3,7 @@ import 'jquery-ui/ui/effect.js';
 import 'jquery-ui/ui/effects/effect-shake.js';
 import 'jquery-ui/ui/widgets/dialog';
 
-import '../sass/ledenmemory.scss';
+import '../../sass/ledenmemory.scss';
 
 $(() => {
 	let first = true;
