+++ resources/assets/js/lib/fotoalbum/main.ts
@@ -18,7 +18,7 @@ declare module 'jgallery/types/gallery/parameters' {
 	}
 }
 
-const loadFotoAlbum = async () => {
+export const loadFotoAlbum = async () => {
 	const albums = document.querySelectorAll<HTMLElement>('.fotoalbum');
 	for (const album of albums) {
 		const {isLoggedIn, magAanpassen, root, fotos} = album.dataset;
@@ -48,7 +48,3 @@ const loadFotoAlbum = async () => {
 		}).getElement());
 	}
 };
-
-// loadFotoAlbum moet async zijn.
-// noinspection JSIgnoredPromiseFromCall
-loadFotoAlbum();
