+++ resources/assets/js/lib/fotoalbum/with-full-size-preview.ts
