+++ resources/assets/js/lib/fotoalbum/with-fullscreen-button.ts
