+++ resources/assets/js/lib/fotoalbum/with-hotkeys.ts
