+++ resources/assets/js/lib/grafiek.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import Chart, {ChartData, ChartOptions} from 'chart.js';
 import palette from 'google-palette';
-import {formatBedrag, html} from './lib/util';
+import {formatBedrag, html} from './util';
 
 function createCanvas(parent: HTMLElement) {
 	const canvas = html`<canvas style="width: 100%; height: 100%"/>` as HTMLCanvasElement;
@@ -121,49 +121,53 @@ export function initDeelnamegrafiek(el: HTMLElement) {
 	return new Chart(createCanvas(el), {data, type: 'bar', options});
 }
 
-Chart.defaults.NegativeTransparentLine = Chart.helpers.clone(Chart.defaults.line);
-Chart.controllers.NegativeTransparentLine = Chart.controllers.line.extend({
-	update() {
-		if (this.chart.data.datasets.length) {
-			// get the min and max values
-			const min = this.chart.data.datasets[0].data
-				.reduce((mininum: number, p: any) => p.y < mininum ? p.y : mininum, this.chart.data.datasets[0].data[0].y);
+function createNegativetransparentLineChartController() {
+	if (Chart.defaults.NegativeTransparentLine) { return; }
 
-			if (min >= 0) {
-				this.chart.data.datasets[0].borderColor = 'green';
-				return Chart.controllers.line.prototype.update.apply(this, arguments);
+	Chart.defaults.NegativeTransparentLine = Chart.helpers.clone(Chart.defaults.line);
+	Chart.controllers.NegativeTransparentLine = Chart.controllers.line.extend({
+		update() {
+			if (this.chart.data.datasets.length) {
+				// get the min and max values
+				const min = this.chart.data.datasets[0].data
+					.reduce((mininum: number, p: any) => p.y < mininum ? p.y : mininum, this.chart.data.datasets[0].data[0].y);
+
+				if (min >= 0) {
+					this.chart.data.datasets[0].borderColor = 'green';
+					return Chart.controllers.line.prototype.update.apply(this, arguments);
+				}
+
+				const max = this.chart.data.datasets[0].data
+					.reduce((maximum: number, p: any) => p.y > maximum ? p.y : maximum, this.chart.data.datasets[0].data[0].y);
+
+				if (max <= 0) {
+					this.chart.data.datasets[0].borderColor = 'red';
+					return Chart.controllers.line.prototype.update.apply(this, arguments);
+				}
+
+				const yScale = this.getScaleForId(this.getMeta().yAxisID);
+
+				// figure out the pixels for these and the value 0
+				const top = yScale.getPixelForValue(max);
+				const zero = yScale.getPixelForValue(0);
+				const bottom = yScale.getPixelForValue(min);
+
+				// build a gradient that switches color at the 0 point
+				const context = this.chart.chart.ctx;
+				const gradient = context.createLinearGradient(0, top, 0, bottom);
+				const ratio = Math.min((zero - top) / (bottom - top), 1);
+				gradient.addColorStop(0, 'green');
+				gradient.addColorStop(ratio, 'green');
+				gradient.addColorStop(ratio, 'red');
+				gradient.addColorStop(1, 'red');
+				this.chart.data.datasets[0].borderColor = gradient;
 			}
 
-			const max = this.chart.data.datasets[0].data
-				.reduce((maximum: number, p: any) => p.y > maximum ? p.y : maximum, this.chart.data.datasets[0].data[0].y);
-
-			if (max <= 0) {
-				this.chart.data.datasets[0].borderColor = 'red';
-				return Chart.controllers.line.prototype.update.apply(this, arguments);
-			}
-
-			const yScale = this.getScaleForId(this.getMeta().yAxisID);
-
-			// figure out the pixels for these and the value 0
-			const top = yScale.getPixelForValue(max);
-			const zero = yScale.getPixelForValue(0);
-			const bottom = yScale.getPixelForValue(min);
-
-			// build a gradient that switches color at the 0 point
-			const context = this.chart.chart.ctx;
-			const gradient = context.createLinearGradient(0, top, 0, bottom);
-			const ratio = Math.min((zero - top) / (bottom - top), 1);
-			gradient.addColorStop(0, 'green');
-			gradient.addColorStop(ratio, 'green');
-			gradient.addColorStop(ratio, 'red');
-			gradient.addColorStop(1, 'red');
-			this.chart.data.datasets[0].borderColor = gradient;
-		}
-
-		// noinspection JSPotentiallyInvalidConstructorUsage
-		return Chart.controllers.line.prototype.update.apply(this, arguments);
-	},
-});
+			// noinspection JSPotentiallyInvalidConstructorUsage
+			return Chart.controllers.line.prototype.update.apply(this, arguments);
+		},
+	});
+}
 
 export function initSaldoGrafiek(el: HTMLElement) {
 	const closed = el.dataset.closed === 'true';
@@ -193,6 +197,8 @@ export function initSaldoGrafiek(el: HTMLElement) {
 			},
 		},
 	};
+
+	createNegativetransparentLineChartController();
 
 	const chart = new Chart(createCanvas(el), {data: {}, type: 'NegativeTransparentLine', options});
