+++ resources/assets/js/lib/maalcie.ts
@@ -124,3 +124,25 @@ export function takenRuilen(e: Event) {
 	}
 	ajaxRequest('POST', elmnt.getAttribute('href')!, 'uid=' + attr, source, domUpdate, alert);
 }
+
+let lastSelectedId: string;
+
+export function takenSelectRange(e: KeyboardEvent) {
+	let withinRange = false;
+	$('#maalcie-tabel').find('tbody tr td a input[name="' + $(e.target!).attr('name') + '"]:visible').each(function () {
+		const thisId = $(this).attr('id');
+		if (thisId === lastSelectedId) {
+			withinRange = !withinRange;
+		}
+		if (thisId === (e.target as Element).id) {
+			withinRange = !withinRange;
+			const check = $(this).prop('checked');
+			setTimeout(() => { // workaround e.preventDefault()
+				$('#' + thisId).prop('checked', check);
+			}, 50);
+		} else if (e.shiftKey && withinRange) {
+			$(this).prop('checked', true);
+		}
+	});
+	lastSelectedId = (e.target as Element).id;
+}
