+++ resources/assets/js/page/fotoalbum.ts
@@ -1,0 +1,5 @@
+import {loadFotoAlbum} from '../lib/fotoalbum/main';
+
+// loadFotoAlbum moet async zijn.
+// noinspection JSIgnoredPromiseFromCall
+loadFotoAlbum();
