+++ resources/assets/js/page/maalcie.ts
@@ -1,0 +1,13 @@
+/**
+ * maalcie.js  |  P.W.G. Brussee (brussee@live.nl)
+ */
+import $ from 'jquery';
+import {takenMagRuilen, takenRuilen} from '../lib/maalcie';
+
+$(() => {
+	$('a.ruilen').each(function () {
+		$(this).removeClass('ruilen');
+		$(this).on('dragover', takenMagRuilen);
+		$(this).on('drop', takenRuilen);
+	});
+});
