+++ resources/assets/sass/common.scss
@@ -10,7 +10,6 @@ $fa-font-path: "~@fortawesome/fontawesome-free/webfonts";
 @import "~lightbox2/dist/css/lightbox.css";
 @import "~jquery-ui/themes/base/core.css";
 @import "~jquery-ui/themes/base/theme.css";
-@import "~jquery-ui/themes/base/tooltip.css";
 @import "~jquery-ui/themes/base/datepicker.css";
 @import "~jquery-ui/themes/base/slider.css";
 @import "~jquery-ui-timepicker-addon/src/jquery-ui-timepicker-addon.css";
