+++ resources/assets/sass/extern.scss
@@ -3,7 +3,6 @@
 @import "~@fortawesome/fontawesome-free/scss/regular";
 @import "~@fortawesome/fontawesome-free/scss/brands";
 @import "~@fortawesome/fontawesome-free/scss/fontawesome";
-@import "~jquery-ui/themes/base/tooltip.css";
 
 @import "fonts";
 @import "extern/variables";
