+++ webpack.config.ts
@@ -8,21 +8,21 @@ const config: (env: string, argv: any) => webpack.Configuration = (env, argv) =>
 	mode: 'development',
 	context: contextPath,
 	entry: {
-		'app': './js/app.ts',
-		'ledenmemory': './js/ledenmemory.ts',
-		'fxclouds': './js/effect/fxclouds.ts',
-		'fxsneeuw': './js/effect/fxsneeuw.ts',
-		'fxonontdekt': './js/effect/fxonontdekt.ts',
-		'fxtrein': './js/effect/fxtrein.ts',
-		'fxraket': './js/effect/fxraket.ts',
-		'fxminion': './js/effect/minion.ts',
-		'fxclippy': './js/effect/fxclippy.ts',
-		'extern': ['./js/extern.ts', './sass/extern.scss'],
+		'app': './js/entry/app.ts',
+		'ledenmemory': './js/entry/ledenmemory.ts',
+		'fxclouds': './js/entry/fxclouds.ts',
+		'fxsneeuw': './js/entry/fxsneeuw.ts',
+		'fxonontdekt': './js/entry/fxonontdekt.ts',
+		'fxtrein': './js/entry/fxtrein.ts',
+		'fxraket': './js/entry/fxraket.ts',
+		'fxminion': './js/entry/fxminion.ts',
+		'fxclippy': './js/entry/fxclippy.ts',
+		'extern': ['./js/entry/extern.ts', './sass/extern.scss'],
 		'bredeletters': './sass/bredeletters.scss',
 		'common': './sass/common.scss',
 		'extern-forum': './sass/extern-forum.scss',
 		'extern-fotoalbum': './sass/extern-fotoalbum.scss',
-		'extern-owee': ['./js/extern-owee.ts', './sass/extern-owee.scss'],
+		'extern-owee': ['./js/entry/extern-owee.ts', './sass/extern-owee.scss'],
 		'maaltijdlijst': './sass/maaltijdlijst.scss',
 		'thema-civitasia': './sass/thema/civitasia.scss',
 		'thema-dies': './sass/thema/dies.scss',
